+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.1.0
+
+- Files can be renamed directly in the modal
+- Pressing ENTER submits
+
 ## v1.0.0
 
 - Initial release